On 32-bit systems, number of seconds in a GPS epoch may overflow
On a gcc 8.3 on a 32-bit system,
Size of int: 4 bytes
Size of long int: 4 bytes
Size of long long int: 8 bytes
Size of float: 4 bytes
Size of double: 8 bytes
Size of long double: 12 bytes
Size of char: 1 byte

Explicitly cast to (long double) to prevent overflow of the
seconds in a 13-bit week number

+++ ntpd/ntp_scanner.c
@@ -35,7 +35,7 @@
 #include "ntp_keyword.h"
 
 /* used to implement g and G suffixes for numeric literals in fudge offset declarations */
-#define SECONDS_IN_WEEK	(7L * 24L * 60L * 60L)
+#define SECONDS_IN_WEEK	(long double)(7L * 24L * 60L * 60L) /* Cast for 32bit */
 #define GPS_ERA_10BIT	(1024L * SECONDS_IN_WEEK)
 #define GPS_ERA_13BIT	(8192L * SECONDS_IN_WEEK)
 #define ERA_SUFFIX(c)	((c) == 'g' || (c) == 'G')
